Create the minio permissions in CI
